style(path): move package from config to config.db
