+++ src/main/java/com/e2i/wemeet/web/config/db/RedisConfig.java
@@ -1,4 +1,4 @@
-package com.e2i.wemeet.web.config;
+package com.e2i.wemeet.web.config.db;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
